Moved dictionary search to utils
The function which searches the outputs for non-built-in data types is now in the csv_utils.py file, as both the extractor and reducer tests factories use it.
